data_eng_ekz/requirements.txt
pandas==2.3.0
scikit_learn==1.7.0
apache-airflow==3.0.2
